+++ rewrite-benchmarks/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     jmh("org.openjdk.jmh:jmh-core:latest.release")
     jmh("org.openjdk.jol:jol-core:latest.release")
     jmh("io.github.fastfilter:fastfilter:latest.release")
+    jmh("org.xerial.snappy:snappy-java:1.1.10.7")
 
     // Nebula doesn't like having jmhAnnotationProcessor without jmh so we just add it twice.
     jmh("org.openjdk.jmh:jmh-generator-annprocess:latest.release")
