+++ rewrite-benchmarks/src/jmh/java/org/openrewrite/benchmarks/java/JavaCompilationUnitState.java
@@ -24,10 +24,13 @@ import org.openrewrite.LargeSourceSet;
 import org.openrewrite.SourceFile;
 import org.openrewrite.internal.InMemoryLargeSourceSet;
 import org.openrewrite.java.JavaParser;
-import org.openrewrite.java.internal.AdaptiveRadixJavaTypeCache;
 import org.openrewrite.java.internal.JavaTypeCache;
+import org.xerial.snappy.Snappy;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.net.URISyntaxException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
@@ -36,17 +39,20 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+@SuppressWarnings("NotNullFieldNotInitialized")
 @State(Scope.Benchmark)
 public class JavaCompilationUnitState {
     JavaParser.Builder<? extends JavaParser, ?> javaParser;
     List<SourceFile> sourceFiles;
     List<Path> inputs;
-    JavaTypeCache snappyTypeCache;
-    AdaptiveRadixJavaTypeCache radixMapTypeCache;
+    SnappyJavaTypeCache snappyTypeCache;
+    JavaTypeCache radixMapTypeCache;
     MapJavaTypeCache typeCache;
 
     public static void main(String[] args) throws URISyntaxException {
-        new JavaCompilationUnitState().setup();
+        JavaCompilationUnitState javaCompiler = new JavaCompilationUnitState();
+        javaCompiler.setup();
+        javaCompiler.printMemory();
     }
 
     @Setup(Level.Trial)
@@ -57,6 +63,7 @@ public class JavaCompilationUnitState {
         inputs = Arrays.asList(
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/lang/Nullable.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/lang/NullUtils.java"),
+                rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/AdaptiveRadixTree.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/MetricsHelper.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/ListUtils.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/internal/PropertyPlaceholderHelper.java"),
@@ -70,6 +77,7 @@ public class JavaCompilationUnitState {
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/style/Style.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/config/DeclarativeNamedStyles.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/style/NamedStyles.java"),
+                rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/Incubating.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/Option.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/config/OptionDescriptor.java"),
                 rewriteRoot.resolve("rewrite-core/src/main/java/org/openrewrite/config/RecipeDescriptor.java"),
@@ -99,12 +107,12 @@ public class JavaCompilationUnitState {
                 .parse(inputs, null, new InMemoryExecutionContext())
                 .collect(Collectors.toList());
 
-        radixMapTypeCache = new AdaptiveRadixJavaTypeCache();
+        radixMapTypeCache = new JavaTypeCache();
         for (Map.Entry<String, Object> entry : typeCache.map().entrySet()) {
             radixMapTypeCache.put(entry.getKey(), entry.getValue());
         }
 
-        snappyTypeCache = new JavaTypeCache();
+        snappyTypeCache = new SnappyJavaTypeCache();
         for (Map.Entry<String, Object> entry : typeCache.map().entrySet()) {
             snappyTypeCache.put(entry.getKey(), entry.getValue());
         }
@@ -115,6 +123,8 @@ public class JavaCompilationUnitState {
         System.out.printf("Retained AdaptiveRadixTree size: %10d bytes\n", retainedSize);
         retainedSize = GraphLayout.parseInstance(snappyTypeCache).totalSize();
         System.out.printf("Retained Snappy size:            %10d bytes\n", retainedSize);
+        retainedSize = GraphLayout.parseInstance(typeCache).totalSize();
+        System.out.printf("Retained HashMap size:           %10d bytes\n", retainedSize);
     }
 
     @TearDown(Level.Trial)
@@ -155,13 +165,64 @@ public class JavaCompilationUnitState {
         }
 
         @Override
-        public int size() {
-            return typeCache.size();
+        public MapJavaTypeCache clone() {
+            MapJavaTypeCache clone = (MapJavaTypeCache) super.clone();
+            clone.typeCache = new HashMap<>(this.typeCache);
+            return clone;
+        }
+    }
+
+    static class SnappyJavaTypeCache extends JavaTypeCache {
+
+        // empirical value: below this size, the compressed key is larger or only slightly smaller
+        // although also note that a String object has a 24 bytes overhead vs. the 16 bytes of a BytesKey object
+        public static final int COMPRESSION_THRESHOLD = 50;
+
+        @SuppressWarnings("ClassCanBeRecord")
+        private static class BytesKey {
+            private final byte[] data;
+            BytesKey(byte[] data) {
+                this.data = data;
+            }
+        }
+
+        Map<Object, Object> typeCache = new HashMap<>();
+
+        @Override
+        public <T> @Nullable T get(String signature) {
+            //noinspection unchecked
+            return (T) typeCache.get(key(signature));
         }
 
         @Override
-        public MapJavaTypeCache clone() {
-            MapJavaTypeCache clone = (MapJavaTypeCache) super.clone();
+        public void put(String signature, Object o) {
+            typeCache.put(key(signature), o);
+        }
+
+        private static boolean snappyUsable = true;
+
+        private Object key(String signature) {
+            if (signature.length() > COMPRESSION_THRESHOLD && snappyUsable) {
+                try {
+                    return new BytesKey(Snappy.compress(signature.getBytes(StandardCharsets.UTF_8)));
+                } catch (IOException e) {
+                    throw new UncheckedIOException(e);
+                } catch (NoClassDefFoundError e) {
+                    // Some systems fail to load Snappy native components, so fall back to not compressing
+                    snappyUsable = false;
+                }
+            }
+            return signature;
+        }
+
+        @Override
+        public void clear() {
+            typeCache.clear();
+        }
+
+        @Override
+        public SnappyJavaTypeCache clone() {
+            SnappyJavaTypeCache clone = (SnappyJavaTypeCache) super.clone();
             clone.typeCache = new HashMap<>(this.typeCache);
             return clone;
         }
