+++ rewrite-benchmarks/src/jmh/java/org/openrewrite/benchmarks/java/JavaTypeCacheBenchmark.java
@@ -22,16 +22,15 @@ import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
-import org.openrewrite.java.internal.AdaptiveRadixJavaTypeCache;
 import org.openrewrite.java.internal.JavaTypeCache;
 
 import java.net.URISyntaxException;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-@Fork(1)
-@Measurement(iterations = 2)
-@Warmup(iterations = 2)
+@Fork(value = 1)
+@Measurement(iterations = 3, time = 5)
+@Warmup(iterations = 3, time = 5)
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.SECONDS)
 @Threads(4)
@@ -39,6 +38,22 @@ public class JavaTypeCacheBenchmark {
 
     @Benchmark
     public void writeSnappy(JavaCompilationUnitState state, Blackhole bh) {
+        JavaTypeCache typeCache = new JavaCompilationUnitState.SnappyJavaTypeCache();
+        for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
+            typeCache.put(entry.getKey(), entry.getValue());
+        }
+    }
+
+//    @Benchmark
+    public void writeHash(JavaCompilationUnitState state, Blackhole bh) {
+        JavaTypeCache typeCache = new JavaCompilationUnitState.MapJavaTypeCache();
+        for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
+            typeCache.put(new String(entry.getKey()), entry.getValue());
+        }
+    }
+
+    @Benchmark
+    public void writeAdaptiveRadix(JavaCompilationUnitState state, Blackhole bh) {
         JavaTypeCache typeCache = new JavaTypeCache();
         for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
             typeCache.put(entry.getKey(), entry.getValue());
@@ -46,8 +61,9 @@ public class JavaTypeCacheBenchmark {
     }
 
     @Benchmark
-    public void writeAdaptiveRadix(JavaCompilationUnitState state, Blackhole bh) {
-        AdaptiveRadixJavaTypeCache typeCache = new AdaptiveRadixJavaTypeCache();
+    @Fork(value = 1, jvmArgsAppend = {"--add-opens", "java.base/java.lang=ALL-UNNAMED"})
+    public void writeAdaptiveRadixAddOpens(JavaCompilationUnitState state, Blackhole bh) {
+        JavaTypeCache typeCache = new JavaTypeCache();
         for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
             typeCache.put(entry.getKey(), entry.getValue());
         }
@@ -60,8 +76,23 @@ public class JavaTypeCacheBenchmark {
         }
     }
 
+//    @Benchmark
+    public void readHash(JavaCompilationUnitState state, Blackhole bh) {
+        for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
+            bh.consume(state.typeCache.get(entry.getKey()));
+        }
+    }
+
     @Benchmark
     public void readAdaptiveRadix(JavaCompilationUnitState state, Blackhole bh) {
+        for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
+            bh.consume(state.radixMapTypeCache.get(entry.getKey()));
+        }
+    }
+
+    @Benchmark
+    @Fork(value = 1, jvmArgsAppend = {"--add-opens", "java.base/java.lang=ALL-UNNAMED"})
+    public void readAdaptiveRadixAddOpens(JavaCompilationUnitState state, Blackhole bh) {
         for (Map.Entry<String, Object> entry : state.typeCache.map().entrySet()) {
             bh.consume(state.radixMapTypeCache.get(entry.getKey()));
         }
