+++ rewrite-java/build.gradle.kts
@@ -50,8 +50,6 @@ dependencies {
     implementation("org.apache.commons:commons-text:latest.release")
     implementation("io.github.classgraph:classgraph:latest.release")
 
-    implementation("org.xerial.snappy:snappy-java:1.1.10.+")
-
     api("com.fasterxml.jackson.core:jackson-annotations")
 
     // these are required for now so that `ChangeType` and `ChangePackage` can use the `Reference` trait
