+++ rewrite-java/src/main/java/org/openrewrite/java/internal/JavaTypeCache.java
@@ -15,72 +15,99 @@
  */
 package org.openrewrite.java.internal;
 
-import lombok.Value;
 import org.jspecify.annotations.Nullable;
-import org.xerial.snappy.Snappy;
+import org.openrewrite.internal.AdaptiveRadixTree;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
+import java.lang.reflect.Field;
 import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
-import java.util.Map;
 
 public class JavaTypeCache implements Cloneable {
 
-    // empirical value: below this size, the compressed key is larger or only slightly smaller
-    // although also note that a String object has a 24 bytes overhead vs. the 16 bytes of a BytesKey object
-    public static final int COMPRESSION_THRESHOLD = 50;
-
-    @SuppressWarnings("ClassCanBeRecord")
-    @Value
-    private static class BytesKey {
-        byte[] data;
-    }
-
-    Map<Object, Object> typeCache = new HashMap<>();
+    AdaptiveRadixTree<Object> typeCache = new AdaptiveRadixTree<>();
 
     public <T> @Nullable T get(String signature) {
         //noinspection unchecked
-        return (T) typeCache.get(key(signature));
+        return (T) typeCache.search(getKeyBytes(signature));
     }
 
     public void put(String signature, Object o) {
-        typeCache.put(key(signature), o);
-    }
-
-    @Nullable
-    private static boolean snappyUsable = true;
-
-    private Object key(String signature) {
-        if (signature.length() > COMPRESSION_THRESHOLD && snappyUsable) {
-            try {
-                return new BytesKey(Snappy.compress(signature.getBytes(StandardCharsets.UTF_8)));
-            } catch (IOException e) {
-                throw new UncheckedIOException(e);
-            } catch (NoClassDefFoundError e) {
-                // Some systems fail to load Snappy native components, so fall back to not compressing
-                snappyUsable = false;
-            }
-        }
-        return signature;
+        typeCache.insert(getKeyBytes(signature), o);
     }
 
     public void clear() {
         typeCache.clear();
     }
 
-    public int size() {
-        return typeCache.size();
-    }
-
     @Override
     public JavaTypeCache clone() {
         try {
             JavaTypeCache clone = (JavaTypeCache) super.clone();
-            clone.typeCache = new HashMap<>(this.typeCache);
+            clone.typeCache = this.typeCache.copy();
             return clone;
         } catch (CloneNotSupportedException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    private static final @Nullable Field STRING_VALUE;
+    private static final @Nullable Field STRING_CODER;
+    private static final boolean USE_REFLECTION;
+
+    static {
+        Field value;
+        Field coder;
+        boolean hasCompactStrings = false;
+
+        try {
+            // requires: --add-opens java.base/java.lang=ALL-UNNAMED
+            value = String.class.getDeclaredField("value");
+            value.setAccessible(true);
+
+            try {
+                coder = String.class.getDeclaredField("coder");
+                coder.setAccessible(true);
+                Field compactStrings = String.class.getDeclaredField("COMPACT_STRINGS");
+                compactStrings.setAccessible(true);
+                hasCompactStrings = compactStrings.getBoolean(null);
+            } catch (NoSuchFieldException e) {
+                // Java 8 - field doesn't exist
+                coder = null;
+            }
+        } catch (Exception e) {
+            value = null;
+            coder = null;
+        }
+
+        STRING_VALUE = value;
+        STRING_CODER = coder;
+        USE_REFLECTION = STRING_VALUE != null && STRING_CODER != null && hasCompactStrings;
+    }
+
+    /**
+     * For ASCII and Latin-1 strings this operation is allocation-free.
+     */
+    static byte[] getKeyBytes(String s) {
+        // Try to get internal representation first
+        if (USE_REFLECTION) {
+            try {
+                //noinspection DataFlowIssue
+                byte[] bytes = (byte[]) STRING_VALUE.get(s);
+                //noinspection DataFlowIssue
+                byte coder = (byte) STRING_CODER.get(s);
+                if (coder == 0) {
+                    // Latin1, use directly
+                    return bytes;
+                } else {
+                    // UTF-8: append NUL byte to avoid collisions
+                    byte[] prefixed = new byte[bytes.length + 1];
+                    System.arraycopy(bytes, 0, prefixed, 0, bytes.length);
+                    prefixed[bytes.length] = 0;
+                    return prefixed;
+                }
+            } catch (Exception ignored) {
+            }
+        }
+
+        return s.getBytes(StandardCharsets.UTF_8);
     }
 }
